tracker/requirements.txt
requests
beautifulsoup4
PyYAML
tenacity
supabase
dotenv
getuseragent
python-telegram-bot==22.4
httpx==0.28.1
selenium
